separate logics
